Added functions to check and adjust value of Ace
Value of Ace is now dependent on the total of the player/dealer hand. If during play the hand total exceeds 21, a function runs to check if the hand has an Ace card valued at 11. If this is found to be true, the hand total is reduced by 10 so that the Ace counts as a value of 1 and the game continues.